feat: handle the form submit for register and login
Create form data object that will be sent to the backend route to be
handled by the controller. If get a response then reset the form and
change page type to login (register). Or if logging in dispatch the
setLogin function to set the user and token state based on the response
set back from the backend controller